[Texture loading & Transform rotation]
Ported over the Texture class and tested it to load an image file from romfs. Works fine.

Tried to rotate and found an issue that glm::quat is not initialised to identity quaternion by default like in the PC implementation.
Simply had to pass the identity values myself in transform's constructor, fixed it up.

Added a new shader for texture usage in romfs.