[Entity Component Architecture - The beginning]
Began porting over the Entity component architecture code, starting with the allocator and the base Object.
However, a few errors occured with forward decleration and const typing so I had to modify it a bit to get it working.
Ideally, some form of const typing will help protect the interface, but for now I'll deal with it when the time comes.

Tested and working, next stage is to port over the component and gameobject classes, which will let me move onto the rendering components and base services/systems.