[GameObject + Component Classes]
The behaviour of these classes seems to be behaving oddly when it comes to the instancing of the GameObject class and it cleaning up it's dangling references.
The PC counterpart doesn't seem to bring up the same issues, so I'm assuming the compiler might be correcting some things for me in comparison to the compiler in devkitpro.

So, I'll need to investigate far more thoroughly or redesign how it works and detach from trying 100% to allow Unity-esque behaviour with GameObject container classes.
I could also decouple the instance and container from the 1 class itself(GameObject) but I quite like the easy interfacing(for those coming from Unity, c veterans will probably be appalled by the existance of this class)

Instantiation and Destruction tested, destroying the original and then trailing to the copy seems to have a leak, but destroying the copy and trailing to the original causes no leaks.
My reference copy logic is messed up somewhere so I'll look for that next. After which, the component and render services can be ported over to start testing update and render components.

[PrimitiveShapes Bug]
Forgot to assign colour values to the majority of the shapes made in the namespace.