[GameObject, Component, UpdateableComponent + ComponentQueue]
Ported over the UpdateableComponent and the queue responsible for keeping track of them.
Updated the header guards since they were clashing with my project folder from the PC version.
Found the bug for the name issue with the gameobjects. Forgot I need to use "->" to access the unique instance, and not use the "." operator as that accesses the container of the instance.