[RenderComponent + Queue port, and cleanups]
Ported over the components and classes responsible for rendering.
Render components haven't been tested yet, but once MeshRendererComponent and UIRendererComponent are ported it will be quick to test.

Cleaned up some of the warnings that kept cropping up when I compiled (which just made it annoying to sift through for the real errors).
Added some markup to the readme to make it easier to read.