[MeshRenderComponent & FPSMovementComponent]
Added in the MeshRenderComponent and FPSMovementComponent to test if they can be added to a game object and it just works.
Found a logic bug with Material when it's default constructor is called and then loadShader is called on it. The shader pointer can have garbage data which causes a crash when trying to be deleted.

Some cleanup for the components needs to be considered since friend class necessity is a crutch and is causing a lot of design issues. Might just make the members needed public, or create a base class/struct that the queues are friends to but the components have a member of.