[MeshRenderer Object + Material Object]
Brought across the MeshRenderer component as its own object to better organise rendering and such.
However, it has some initialisation issues if it is constructed before OpenGLLoader is initialised.
Refactoring may be needed, however this is only and issue if it is a global variable in main.cpp.
Still it's a good idea to see if I can refactor it better to not just cause crashes by existing as a member of something.
One way could be to make it check if OpenGL is initialised and force it to init, but then it becomes a bad runtime object.

Also fixed the PrimitiveShapes definition file... thought I migrated all the necessary changes from the other branch.