[Scene implementation + main.cpp cleanup]
Wanted to remove the clutter from the main file, and start focusing on initialisation order.

Made a scene class to be inherited to just keep a barebones structure for a scene to start being able to make different files to test features and whatnot.
However I got it to crash once and cannot get it to happen again. I'm almost certain it has to do with MeshRenderer, Material, Shader, or all of them together.

I think I need to make a GameObject class to keep all these things together, with light component architecture (no backend system/engine, just an object with component extensibility.).
That way I can control the destruction of objects with a factory, or I can make a struct and simply make a factory method to better handle the creation and destruction of these things.
Minimum is best, as main focus of this branch is to refine the barebones things like Input, Timer, etc.