[Cross Platform Refactor - Mingw64]
Modified Glad to use a version without 'Generate a loader' ticked on the website. This let glad cross compile.

Project was heavily modified to allow compiling on both Switch and Windows. This was done to help speed testing and development of features. As well as being an exercise in setting mingw64 and learning how to use the GDB on the command line to debug programs.

--- Framework Features + Modifications ---

Added obj loader using TinyObjLoader.

Rendering model made from MagicaVoxel.

Found main loop was unclamped and was running as fast as possible. Crude implementation of limiting framerate via sleep_for. Helps conserver battery life and keep rendering results and movements looking similar between the Switch and Windows.

Shaders and Materials are cached and don't have repeated compiling. Need to rewrite them so they are copyable and destructable without harming other shaders or materials that reference the same program.

Discovered a lot of bugs and initialisation errors due to compiling and debugging it on Windows.