[Fixing the Branch Merge...]
Removed unnecessary constructors and destructors from a few classes.

Fixed up a few merge issues (more to come).

Added references to libraries used in the Readme file.

Moved old ECA prototyping code from master branch into a separate folder to not break the current version of the project.

--- Details ---

A branch merge was required due to a huge amount of major changes on the base-framework branch compared to the master branch.

Led to a while of fiddling around and merging. However some old code meant for the ECA style early on in the project was merged into the base-framework automatically by accident.

A thorough comparison will need to be made to make sure that the base-framework Cross Compile commit changes are the most up to date changes for the framework code. Worst case, rebase or delete and override project folder to make sure.